moved map claims to its own file
